Fix data loading bug, extend page config, small improvements
When loading existing data from file, the app would not take in new user data. The reason is that once a file was loaded, it would be re-read at every UI interaction (since Streamlit runs top-to-bottom), and the entered data would be overwritten with the file data.

I fixed two aspects of this:
1. Loaded data EXTEND other data rather than REPLACING them
2. The loading code runs only iff file is uploaded AND button is pressed.